Completing the rest of the CV page
The previous commit did work and fix the CSS not loading (it was the / lol). Now finished everything else, and can look to add more things, or make it look cooler.

(While writing this I realised I haven't added a link to the certificate so I will do that next commit).